--- a/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
+++ b/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
@@ -2386,7 +2386,7 @@ void MacroAssembler::store_heap_oop_null(Address dst) {
 }
 
 int MacroAssembler::corrected_idivl(Register result, Register rs1, Register rs2,
-                                    bool want_remainder)
+                                    bool want_remainder, bool is_signed)
 {
   // Full implementation of Java idiv and irem.  The function
   // returns the (pc) offset of the div instruction - may be needed
@@ -2402,7 +2402,11 @@ int MacroAssembler::corrected_idivl(Register result, Register rs1, Register rs2,
 
   int idivl_offset = offset();
   if (!want_remainder) {
-    divw(result, rs1, rs2);
+    if (is_signed) {
+      divw(result, rs1, rs2);
+    } else {
+      divuw(result, rs1, rs2);
+    }
   } else {
     remw(result, rs1, rs2); // result = rs1 % rs2;
   }
@@ -2410,7 +2414,7 @@ int MacroAssembler::corrected_idivl(Register result, Register rs1, Register rs2,
 }
 
 int MacroAssembler::corrected_idivq(Register result, Register rs1, Register rs2,
-                                    bool want_remainder)
+                                    bool want_remainder, bool is_signed)
 {
   // Full implementation of Java ldiv and lrem.  The function
   // returns the (pc) offset of the div instruction - may be needed
@@ -2425,7 +2429,20 @@ int MacroAssembler::corrected_idivq(Register result, Register rs1, Register rs2,
 
   int idivq_offset = offset();
   if (!want_remainder) {
-    div(result, rs1, rs2);
+    if (is_signed) {
+      div(result, rs1, rs2);
+    } else {
+      Label Lltz, Ldone;
+      bltz(rs2, Lltz);
+      divu(result, rs1, rs2);
+      j(Ldone);
+      bind(Lltz); // For the algorithm details, check j.l.Long::divideUnsigned
+      sub(result, rs1, rs2);
+      notr(result, result);
+      andr(result, result, rs1);
+      srli(result, result, 63);
+      bind(Ldone);
+    }
   } else {
     rem(result, rs1, rs2); // result = rs1 % rs2;
   }