--- a/src/hotspot/cpu/riscv/riscv.ad
+++ b/src/hotspot/cpu/riscv/riscv.ad
@@ -6673,6 +6673,17 @@ instruct divI(iRegINoSp dst, iRegIorL2I src1, iRegIorL2I src2) %{
   ins_pipe(idiv_reg_reg);
 %}
 
+instruct UdivI(iRegINoSp dst, iRegIorL2I src1, iRegIorL2I src2) %{
+  match(Set dst (UDivI src1 src2));
+  ins_cost(IDIVSI_COST);
+  format %{ "divuw  $dst, $src1, $src2\t#@UdivI"%}
+
+  ins_encode %{
+    __ divuw(as_Register($dst$$reg), as_Register($src1$$reg), as_Register($src2$$reg));
+  %}
+  ins_pipe(idiv_reg_reg);
+%}
+
 instruct signExtract(iRegINoSp dst, iRegIorL2I src1, immI_31 div1, immI_31 div2) %{
   match(Set dst (URShiftI (RShiftI src1 div1) div2));
   ins_cost(ALU_COST);
@@ -6692,6 +6703,38 @@ instruct divL(iRegLNoSp dst, iRegL src1, iRegL src2) %{
   format %{ "div  $dst, $src1, $src2\t#@divL" %}
 
   ins_encode(riscv_enc_div(dst, src1, src2));
+  ins_pipe(ldiv_reg_reg);
+%}
+
+instruct UdivL(iRegLNoSp dst, iRegL src1, iRegL src2) %{
+  match(Set dst (UDivL src1 src2));
+  ins_cost(IDIVDI_COST);
+
+  format %{
+    "bltz $src2, Lltz\t#@UdivL\n\t"
+    "divu  $dst, $src1, $src2\n\t"
+    "j Ldone\n\t"
+    "Lltz:"
+    "sub $dst, $src1, $src2\n\t"
+    "not $dst, $dst\n\t"
+    "and $dst, $dst, $src1\n\t"
+    "srli $dst, $dst, 63\n\t"
+    "Ldone:\t#@UdivL"
+  %}
+
+  ins_encode %{
+    Label Lltz, Ldone;
+    __ bltz(as_Register($src2$$reg), Lltz);
+    __ divu(as_Register($dst$$reg), as_Register($src1$$reg), as_Register($src2$$reg));
+    __ j(Ldone);
+    __ bind(Lltz); // check j.l.Long::divideUnsigned
+    __ sub(as_Register($dst$$reg), as_Register($src1$$reg), as_Register($src2$$reg));
+    __ notr(as_Register($dst$$reg), as_Register($dst$$reg));
+    __ andr(as_Register($dst$$reg), as_Register($dst$$reg), as_Register($src1$$reg));
+    __ srli(as_Register($dst$$reg), as_Register($dst$$reg), 63);
+    __ bind(Ldone);
+  %}
+
   ins_pipe(ldiv_reg_reg);
 %}