Hi,
Can you review the change to add intrinsic for UDivI and UDivL?
Thanks!


## Tests

### Functionality
Run tests successfully found via `grep -nr test/jdk/ -we divideUnsigned` and `grep -nr test/hotspot/ -we divideUnsigned` 

### Performance
( NOTE: there are another 2 related issues: https://bugs.openjdk.org/browse/JDK-8318225, https://bugs.openjdk.org/browse/JDK-8318226, the pr of which will be subseqently sent out after this one finished. )

#### Long
**NOTE: for positive divisor, it's the common case; for negative divisor, it's a rare case**

**Before**
```
LongDivMod.testDivideUnsigned                    1024          mixed  avgt   10  19684.873 ± 21.882  ns/op
LongDivMod.testDivideUnsigned                    1024       positive  avgt   10  28853.041 ±  6.425  ns/op
LongDivMod.testDivideUnsigned                    1024       negative  avgt   10   6367.239 ± 16.011  ns/op
```

**After**
```
LongDivMod.testDivideUnsigned                    1024          mixed  avgt   10  22622.133 ±  7.158  ns/op
LongDivMod.testDivideUnsigned                    1024       positive  avgt   10  15957.272 ±  3.174  ns/op
LongDivMod.testDivideUnsigned                    1024       negative  avgt   10  29499.721 ± 10.404  ns/op
```

#### Integer
**Before**
```
IntegerDivMod.testDivideUnsigned                    1024          mixed  avgt   10  23397.267 ± 36.980  ns/op
IntegerDivMod.testDivideUnsigned                    1024       positive  avgt   10  16792.414 ±  5.869  ns/op
IntegerDivMod.testDivideUnsigned                    1024       negative  avgt   10  30184.357 ± 55.464  ns/op
```

**After**
```
IntegerDivMod.testDivideUnsigned                    1024          mixed  avgt   10  23210.437 ±  4.463  ns/op
IntegerDivMod.testDivideUnsigned                    1024       positive  avgt   10  16622.342 ±  4.047  ns/op
IntegerDivMod.testDivideUnsigned                    1024       negative  avgt   10  30013.414 ± 48.695  ns/op
```


/************ following is just backup: quick path for negative divisor *************/
#### Long
**Before**
```
LongDivMod.testDivideUnsigned                    1024          mixed  avgt   10  19704.317 ± 64.078  ns/op
LongDivMod.testDivideUnsigned                    1024       positive  avgt   10  28856.859 ± 14.901  ns/op
LongDivMod.testDivideUnsigned                    1024       negative  avgt   10   6364.974 ±  2.465  ns/op
```

**After v1**
(This is a simpler version, please check the diff from `After v2` below)
```
LongDivMod.testDivideUnsigned                    1024          mixed  avgt   10  22668.228 ± 74.161  ns/op
LongDivMod.testDivideUnsigned                    1024       positive  avgt   10  15966.320 ± 14.985  ns/op
LongDivMod.testDivideUnsigned                    1024       negative  avgt   10  29518.033 ± 49.056  ns/op
```

**After v2**
(This is the current patch, **This version has a huge regression for negative values!!!**)
```
LongDivMod.testDivideUnsigned                    1024          mixed  avgt   10  11432.738 ±  95.785  ns/op
LongDivMod.testDivideUnsigned                    1024       positive  avgt   10  15969.044 ±  19.492  ns/op
LongDivMod.testDivideUnsigned                    1024       negative  avgt   10   6376.674 ±  16.869  ns/op
```

##### Diff of v1 from v2
```
diff --git a/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp b/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
index b96f7611133..dfb40e171e7 100644
--- a/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
+++ b/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
@@ -2432,16 +2432,7 @@ int MacroAssembler::corrected_idivq(Register result, Register rs1, Register rs2,
     if (is_signed) {
       div(result, rs1, rs2);
     } else {
-      Label Lltz, Ldone;
-      bltz(rs2, Lltz);
       divu(result, rs1, rs2);
-      j(Ldone);
-      bind(Lltz); // For the algorithm details, check j.l.Long::divideUnsigned
-      sub(result, rs1, rs2);
-      notr(result, result);
-      andr(result, result, rs1);
-      srli(result, result, 63);
-      bind(Ldone);
     }
   } else {
     rem(result, rs1, rs2); // result = rs1 % rs2;
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issues
 * [JDK-8318723](https://bugs.openjdk.org/browse/JDK-8318723): RISC-V: C2 UDivL (**Sub-task** - P4)
 * [JDK-8318224](https://bugs.openjdk.org/browse/JDK-8318224): RISC-V: C2 UDivI (**Sub-task** - P4)


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**)
 * [Ludovic Henry](https://openjdk.org/census#luhenry) (@luhenry - Committer)
 * [Andrew Haley](https://openjdk.org/census#aph) (@theRealAph - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/16346/head:pull/16346` \
`$ git checkout pull/16346`

Update a local copy of the PR: \
`$ git checkout pull/16346` \
`$ git pull https://git.openjdk.org/jdk.git pull/16346/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 16346`

View PR using the GUI difftool: \
`$ git pr show -t 16346`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/16346.diff">https://git.openjdk.org/jdk/pull/16346.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/16346#issuecomment-1777488427)