@theRealAph 2023-10-26
> So I tried this on Hifive Unmatched board. Unforunately, JMH test shows some regression for the LongDivMod.testDivideUnsigned `negative` case.

But that case is going to be rare.The larger a number it is, the less common it is. The uniform distribution of this benchmark, in which 0 is as common as 0xb43a61c853a2af20, is grossly unrepresentative of real-world divisors.

In practice, numbers follow some kind of log-normal distribution. Don't fall into the trap of optimizing for a benchmark.