@Hamlin-Li 2023-10-26
`Don't fall into the trap of optimizing for a benchmark.` Agree